Add files via upload
I'm only uploading what is required of my team; links to event-related items will not work, plus database-related functions will not work unless the dbconfig is changed to match whatever database the host is using.  Databases will initially need to be made manually, though with a few lines of code that need can be removed.